dashboard routes, todo: fix provider
