+++ mod-disco/service/flu/lib/core/core.dart
@@ -1,4 +1,3 @@
-export './shared_services/base_model.dart';
 export './routes/paths.dart';
 export './config/env_config.dart';
 export './i18n/mod_disco_localization.dart';
