+++ mod-disco/service/flu/lib/modules/dashboard/views/dashboard_detail_view.dart
@@ -20,7 +20,7 @@ class DashboardDetailView extends StatelessWidget {
     @required this.orgId,
     @required this.projectId,
     @required this.isFullScreen,
-  });
+  }): super(key: key);
 
   @override
   Widget build(BuildContext context) {
