+++ mod-disco/service/flu/lib/modules/main_module.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/widgets.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:mod_disco/core/core.dart';
 import 'package:mod_disco/core/routes/dashboard_guards.dart';
@@ -5,7 +6,9 @@ import 'package:mod_disco/modules/dashboard/views/dashboard_view.dart';
 import 'package:mod_disco/modules/projects/views/proj_view.dart';
 import 'package:mod_disco/modules/survey_project/views/support_role_view.dart';
 import 'package:mod_disco/modules/survey_project/views/survey_project_view.dart';
+import 'package:random_string/random_string.dart';
 import 'package:sys_share_sys_account_service/pkg/guards/guardian_view_model.dart';
+import 'package:sys_share_sys_account_service/sys_share_sys_account_service.dart';
 
 class AdminDashboardModule extends ChildModule {
   final String baseRoute;
@@ -13,39 +16,48 @@ class AdminDashboardModule extends ChildModule {
   AdminDashboardModule({this.baseRoute = '/dashboard'});
 
   @override
-  List<Bind> get binds => [
+  List<Bind> get binds =>
+      [
         Bind.singleton((i) => DashboardPaths(baseRoute)),
-        Bind.lazySingleton((i) => GuardianViewModel())
+        Bind.lazySingleton((i) => GuardianViewModel()),
       ];
 
   @override
-  List<ModularRoute> get routes => [
+  List<ModularRoute> get routes =>
+      [
+
         /// Admin Dashboard Routes
         ChildRoute(
           baseRoute,
-          child: (_, args) => DashboardView(
-            id: args.params['id'] ?? '',
-            orgId: args.params['orgId'] ?? '',
-            routePlaceholder: DashboardPaths(this.baseRoute).dashboardId,
-          ),
+          child: (_, args) =>
+              DashboardView(
+                key: Key(randomString(32)),
+                id: args.params['id'] ?? '',
+                orgId: args.params['orgId'] ?? '',
+                routePlaceholder: DashboardPaths(this.baseRoute).dashboardId,
+              ),
           guards: [DashboardGuard()],
         ),
         ChildRoute(
           '/orgs',
-          child: (_, args) => DashboardView(
-            id: args.params['id'] ?? '',
-            orgId: args.params['orgId'] ?? '',
-            routePlaceholder: DashboardPaths(this.baseRoute).dashboardId,
-          ),
+          child: (_, args) =>
+              DashboardView(
+                key: Key(randomString(32)),
+                id: args.params['id'] ?? '',
+                orgId: args.params['orgId'] ?? '',
+                routePlaceholder: DashboardPaths(this.baseRoute).dashboardId,
+              ),
           guards: [DashboardGuard()],
         ),
         ChildRoute(
           '/orgs/:orgId/:id',
-          child: (_, args) => DashboardView(
-            id: args.params['id'] ?? '',
-            orgId: args.params['orgId'] ?? '',
-            routePlaceholder: DashboardPaths(this.baseRoute).dashboardId,
-          ),
+          child: (_, args) =>
+              DashboardView(
+                key: Key(randomString(32)),
+                id: args.params['id'] ?? '',
+                orgId: args.params['orgId'] ?? '',
+                routePlaceholder: DashboardPaths(this.baseRoute).dashboardId,
+              ),
           guards: [DashboardGuard()],
         ),
       ];
@@ -60,58 +72,74 @@ class MainAppModule extends ChildModule {
     String baseRoute,
     String url,
     String urlNative,
-  })  : this.baseRoute = (baseRoute == '/') ? '' : baseRoute,
+  })
+      : this.baseRoute = (baseRoute == '/') ? '' : baseRoute,
         this.url = url,
         this.urlNative = urlNative;
 
   @override
-  List<Bind> get binds => [
+  List<Bind> get binds =>
+      [
         Bind.singleton((i) => Paths(baseRoute)),
-        Bind.singleton((i) => GuardianViewModel())
+        Bind.singleton((i) => GuardianViewModel()),
       ];
 
   @override
-  List<ModularRoute> get routes => [
+  List<ModularRoute> get routes =>
+      [
         ChildRoute(
-          baseRoute,
-          child: (_, args) => ProjectView(
-            orgs: args.data ?? [],
-            orgId: args.queryParams['orgId'] ?? '',
-            id: args.queryParams['id'] ?? '',
-            routePlaceholder: Paths(this.baseRoute).projectsId,
-          ),
+          '/',
+          child: (_, args) =>
+              ProjectView(
+                key: Key(randomString(32)),
+                orgs: args.data ?? List<Org>.empty(),
+                orgId: args.queryParams['orgId'] ?? '',
+                id: args.queryParams['id'] ?? '',
+                routePlaceholder: Paths(this.baseRoute).projectsId,
+              ),
         ),
         ChildRoute(
           '/subbed/:oid',
-          child: (_, args) => ProjectView(
-            oid: args.params['oid'] ?? '',
-            routePlaceholder: Paths(this.baseRoute).projectsId,
-          ),
+          child: (_, args) =>
+              ProjectView(
+                key: Key(randomString(32)),
+                oid: args.params['oid'] ?? '',
+                orgs: args.data ?? List<Org>.empty(),
+                orgId: args.queryParams['orgId'] ?? '',
+                id: args.queryParams['id'] ?? '',
+                routePlaceholder: Paths(this.baseRoute).projectsId,
+              ),
         ),
         ChildRoute(
           "/projects/:orgId/:id",
-          child: (_, args) => ProjectView(
-            // body: args.data['body'],
-            orgs: args.data ?? [],
-            orgId: args.params['orgId'] ?? '',
-            id: args.params['id'] ?? '',
-            routePlaceholder: Paths(this.baseRoute).projectsId,
-          ),
+          child: (_, args) =>
+              ProjectView(
+                // body: args.data['body'],
+                key: Key(randomString(32)),
+                orgs: args.data ?? List<Org>.empty(),
+                orgId: args.params['orgId'] ?? '',
+                id: args.params['id'] ?? '',
+                routePlaceholder: Paths(this.baseRoute).projectsId,
+              ),
         ),
         ChildRoute(
           "/survey/:id",
-          child: (_, args) => SurveyProjectView(
-            projectId: args.params['id'] ?? '',
-          ),
+          child: (_, args) =>
+              SurveyProjectView(
+                key: Key(randomString(32)),
+                projectId: args.params['id'] ?? '',
+              ),
         ),
         ChildRoute(
           "/support_roles",
-          child: (_, args) => SurveySupportRoleView(
-            project: args.data['project'],
-            surveyUserRequest: args.data['surveyUserRequest'],
-            accountId: args.data['accountId'],
-            surveyProjectList: args.data['surveyProjectList'],
-          ),
+          child: (_, args) =>
+              SurveySupportRoleView(
+                key: Key(randomString(32)),
+                project: args.data['project'],
+                surveyUserRequest: args.data['surveyUserRequest'],
+                accountId: args.data['accountId'],
+                surveyProjectList: args.data['surveyProjectList'],
+              ),
         ),
       ];
