+++ mod-disco/service/flu/lib/modules/projects/view_model/project_detail_view_model.dart
@@ -1,12 +1,11 @@
-import 'package:flutter_modular/flutter_modular.dart';
+import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
-import 'package:mod_disco/core/core.dart';
 import 'package:mod_disco/core/shared_repositories/disco_project_repo.dart';
 import 'package:mod_disco/rpc/v2/mod_disco_models.pb.dart';
 import 'package:sys_share_sys_account_service/pkg/shared_repositories/orgproj_repo.dart';
 import 'package:sys_share_sys_account_service/sys_share_sys_account_service.dart';
 
-class ProjectDetailViewModel extends BaseModel {
+class ProjectDetailViewModel extends ChangeNotifier {
   String projectId;
   Project _proj;
   DiscoProject _discoProject;
