+++ mod-disco/service/flu/lib/modules/projects/views/proj_detail_view.dart
@@ -25,7 +25,7 @@ class ProjectDetailView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ViewModelProvider.withConsumer(
-      disposeViewModel: false,
+      // disposeViewModel: false,
       viewModelBuilder: () => ProjectDetailViewModel(projectId: projectId),
       onModelReady: (ProjectDetailViewModel model) async {
         await model.fetchProjectDetail();
